requirements.txt
anyio==4.9.0
boto3==1.38.9
botocore==1.38.9
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
